requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
async-timeout==5.0.1
billiard==4.2.1
black==25.1.0
blis==1.3.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.1
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
preshed==3.0.10
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
starlette==0.47.2
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.3
